Extract detection case linkage helpers
